Moved utils into a package
- Also added a more generic utils.LimitReadCloser
