+++ server/subscribe.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 
 	jira "github.com/andygrunwald/go-jira"
+	"github.com/mattermost/mattermost-plugin-jira/server/server/utils"
 	"github.com/mattermost/mattermost-server/model"
 	"github.com/pkg/errors"
 )
@@ -333,7 +334,7 @@ func (p *Plugin) hasPermissionToManageSubscription(userId, channelId string) err
 		}
 
 		allowedGroups := strings.Split(cfg.GroupsAllowedToEditJiraSubscriptions, ",")
-		allowedGroups = Map(allowedGroups, strings.TrimSpace)
+		allowedGroups = utils.Map(allowedGroups, strings.TrimSpace)
 		if !inAllowedGroup(groups, allowedGroups) {
 			return errors.New("not in allowed jira user groups")
 		}
